Add Projectionist class for creating a movie -- currently the Projectionist is owend by the keyListener, but it should be moved into StapGraph instead. Hitting Shift+D will now cause Stap to play through the list of functions in the order that they were called.
Change the way the TreeViewer and StapGraph's radial draw mode are integrated. Instead of having the double-click listener highlight the treeview, the radial draw method itself now deals with the Tree View. This prevents us from having to implement treeview integration multiple times.

Change the way StapData handles collapsed nodes. The partOfCollapsedNode variable is now an integer which will contain the value of the collapsed node that the node is a part of, if any, and defaults to the constant NOT_PART_OF_COLLAPSED_NODE defined in StapData.

Other minor organizational changes.

TODO: Move ownership of Projectionist job into StapGraph.