Add support for GDB Tracepoints
[1] Support for experiment disposal

Normally, there is only one case where no experiment is selected: at
startup. However, with GDB Tracepoints, each experiment is tied to a
GDB launch. If Mr. User terminates all his experiments, then none is
selected and the various views (namely the Events View :-) should be
cleared.

This patch adds a new signal (TmfExperimentDisposedSignal) and the
necessary support code. In addition, it provides for some symmetry in
the experiment signals.

[2] Trace re-positionning

The Events view is a actually a window on the trace i.e. for memory
consumption reasons, not all the events are read but only a subset
encompassing the visible area. That subset acts as a cache and is
populated on demand.

For LTTng and other internal tools this is just fine. But for GDB
Tracepoints, populating the cache also sets the current event pointer
to the last event read (not the one that was selected before the read
operation). So the pointer needs to refreshed on cache population
completion. Unfortunately, this breaks the other tracers.

As a temporary fix to have everybody sit on the same version of the
TMF, a hook was added. Today, it is only implemented by GDB
Tracepoints.