oprofile: Modifications to use the linuxtools' path set for the project.
If there's a Linuxtools path defined for the project being profiled,
the plugin will try to run opcontrol as root. If it doesn't work,
user will be prompted to add this opcontrol in the sudoers file with
NOPASSWD. If no path is defined for the project, the plugin will try
to use the opcontrol link/wrapper created by the install scripts.