Add method to get attributes with a wildcard in the path
One can now use getQuarks("x", "*", "z"), and it will return a List
of all attribute quarks whose path matches the "x" and "z" parts.

Not all edge cases have been tested (it fails if the * is a the end),
but proper unit tests will be added shortly. It's also probably
very slow and not yet completely optimized.

Signed-off-by: Alexandre Montplaisir <alexandre.montplaisir@polymtl.ca>