tmf: Add option to specify a resolution to range queries
When we want to populate a view using intervals from the state system,
we don't usually care about the information we can't even show in the
view (for example, for states whose duration are smaller than the length
of one pixel at the current zoom level).

This new variant to queryHistoryRange() adds a 'resolution' parameter,
which is a minimum timestamp between query points. This is only a
minimum: if the current interval is longer than the pixel, we won't
do useless queries in between.

Signed-off-by: Alexandre Montplaisir <alexmonthy@voxpopuli.im>