Add new implementation of CMainTab using ResourceSelectorWidget
Add a new implemetation of CMainTab, called RemoteProxyCMainTab,
which utilizes three instances of ResourceSelectorWidget.  One widget
for selecting the location of the executable to copy from, another
to select its destination, and a third to select the working directory.
Lots of error checking is done, and pains are taken to emit useful
error messages.

Signed-off-by: Corey Ashford <cjashfor@linux.vnet.ibm.com>