Added ability to open perf data files in the Perf Viewer.
In the plug-in.xml file, an editor extension point is included in order
to add a launcher (PerfOpenData.java) for perf data files. In order to
properly report on the the perf file a configuration is needed. This
is done in the createDefaultConfiguration (PerfOpenData.java), which
creates a new default configuration based on the project's name.

In PerfCore.java, it is not necesssary to differentiate between default
events and multiple events, so now events are collected without having
to check this.

Changes were made to the current tests (ModelTest.java) to reflect the
new change, which specifies what the "Default Event" is, namely "cycles".