tmf: Coalesce the state system event request with the indexing
Bring back the previous behaviour of using a signal to trigger
the building of the state history. This time however, we link
the request to the experiment, not the trace, so it gets
coalesced with the request from the indexing. No need for a
separate job then.

It also seems to fix bug #379443, since now events are read
only once.

Signed-off-by: Alexandre Montplaisir <alexmonthy@voxpopuli.im>