lttng: Change external dependencies to imported packages
The external dependencies for ctf and lttng2 (RSE and Antlr) were
set as "required bundles". This means the exact plugin must be
present in the current installation when building the workspace
from within Eclipse.

This commit changes those to Imported Packages, which relaxes the
origin of the package. This allows distribution-installed plugins
to fulfill the build dependency.

Change-Id: Ie20d19959d819fc834dc045ec0520285fae2cf2d
Signed-off-by: Alexandre Montplaisir <alexmonthy@voxpopuli.im>
Reviewed-on: https://git.eclipse.org/r/7021