ctf: Handle absence of stream and event IDs
As per the CTF spec, a trace could define no stream_id field
in its packet headers. In such a case, the parser should assign
it ID 0.

Similarly, if only one event type is present, the event id field
is optional, and 0 should be used if it's not explicitely set.

Partially fixes bug #387039

Change-Id: I38dcdbdefc6c18a16b2ad8fb9ad6135303da208f
Signed-off-by: Alexandre Montplaisir <alexmonthy@voxpopuli.im>
Reviewed-on: https://git.eclipse.org/r/7203
Reviewed-by: Bernd Hufmann <bhufmann@gmail.com>