Updated Enum handling for a CTFTmfEventField
This change provides a way to extract the Enum long value and Enum
value name from the CTF EnumDefinition and store it in a
CtfTmfEventField object. Before this change the CtfTmfEventField
only contained the Enum value name.

For this change, the class Pair<A,B> which was part of the legacy
LTTng core is now integrated in TMF core.

Added JUnit tests for Pair class.

Change-Id: I948d3b9bdde83d5eda7c8397c1b9494e15f7706a
Signed-off-by: Bernd Hufmann <bhufmann@gmail.com>
Reviewed-on: https://git.eclipse.org/r/7326
Reviewed-by: Alexandre Montplaisir <alexmonthy@voxpopuli.im>
IP-Clean: Alexandre Montplaisir <alexmonthy@voxpopuli.im>
Tested-by: Alexandre Montplaisir <alexmonthy@voxpopuli.im>
Reviewed-by: Patrick Tasse <patrick.tasse@gmail.com>
IP-Clean: Patrick Tasse <patrick.tasse@gmail.com>
Tested-by: Patrick Tasse <patrick.tasse@gmail.com>