Do not call the doSave() in a plugin, unless for testing.
The purpose of doSave() is to save the changes made to an
ILaunchConfigurationWorkingCopy to the underlying ILaunchConfiguration.
This is done automatically in most cases. The main callers of doSave()
are methods that set values for a launch configuration (performApply,
setDefaults), but these methods should not perform this functionality.

Change-Id: Id9a0e3c35dac39d2709480f8da1e1f05b7b3fe99
Reviewed-on: https://git.eclipse.org/r/7810
Tested-by: Hudson CI
Reviewed-by: Camilo Bernal <cabernal@redhat.com>
IP-Clean: Camilo Bernal <cabernal@redhat.com>
Tested-by: Camilo Bernal <cabernal@redhat.com>