ctf: Handle traces with unknown event attributes
Instead of blowing up when it sees an attribute is doesn't know,
the parser should just ignore it. This is quite important,
since CTF allows for custom attributes and the parser can't
know in advance about all possible ones.

Ideally, we should also log whenever an unkown attribute is seen.
Unfortunately there is no logging in place in the ctf.core plugins
yet.

Change-Id: I45aba769205060cdcac6935cecc1479b026a5bc0
Signed-off-by: Alexandre Montplaisir <alexmonthy@voxpopuli.im>
Reviewed-on: https://git.eclipse.org/r/7904