tmf: Make TmfStatisticsViewer use the trace's stats provider
Rework the TmfStatisticsViewer to get its information from a
ITmfStatistics object (instead of doing direct event queries
on the trace). That way, the logic to compute the statistics
is now separate from the view, which I think is a good thing.

This removes the need for the ITmfExtraInfo completely. Unit
tests have also been updated to take account for the fact that
the view does not 'count' anymore, it just displays whatever
we send it.

Other changes:
 - Consolidated the setters in TmfStatisticsTree
 - Renamed TmfStatisticsNode.getValue() to .getValues(), to
   better indicate that it is the wrapper object around the
   values that is returned.

Fixes bug 387634.

Change-Id: I1e6b9931007c79cce8c1c93c6c0ddc955cbd0a93
Signed-off-by: Alexandre Montplaisir <alexmonthy@voxpopuli.im>
Reviewed-on: https://git.eclipse.org/r/6986
Tested-by: Hudson CI
Reviewed-by: Patrick Tasse <patrick.tasse@gmail.com>
IP-Clean: Patrick Tasse <patrick.tasse@gmail.com>
Tested-by: Patrick Tasse <patrick.tasse@gmail.com>