Fix default Content column in events table.
The default Content column in TmfEventsTable was being populated using
the event content value's toString if available, or the event content's
toString otherwise.

This was preventing the possibility to show different information in the
Content column and in the raw viewer. The raw viewer is populated using
the event content value's toString.

Therefore if an event's root content had a value defined for use in the
raw viewer, the root content's toString was ignored for use in the table
column, and the column always showed the same data as the raw viewer.

The Content column is now always populated using the event content's
toString. The (optional) event content value's toString is only used by
the raw viewer.

The filter base class implementation is modified to use the same data as
is displayed by the table when filtering on the default Content
column.

Change-Id: Ic976104743eb8850d8176ce232ca84e0574badd3
Reviewed-on: https://git.eclipse.org/r/8570
Tested-by: Hudson CI
Reviewed-by: Bernd Hufmann <bhufmann@gmail.com>
IP-Clean: Bernd Hufmann <bhufmann@gmail.com>
Tested-by: Bernd Hufmann <bhufmann@gmail.com>