Fix deficiencies in our build system.
The javadoc:aggregate goal (maven-javadoc-plugin) executes on modules in
parallel. It also requires that modules resolve their dependencies for
the 'compile' scope. If it is not executed last, there is a risk of
dependency failure when linuxtools is not in the reactor cache.

The javadoc:aggregate goal (maven-javadoc-plugin) also triggers the
'generate-sources' phase of the lifecycle, which causes the 'initialize'
phase to be executed. However, packaging types of 'eclipse-repository'
perform a clean during this phase, so this wipes out our built zip/p2
repo before they can be copied over. We need to preserve our build
directory for later use.

Change-Id: If92b8bf453914e13ac1fa035e07563f0d42e6ce8
Reviewed-on: https://git.eclipse.org/r/8571
Tested-by: Hudson CI
Reviewed-by: Jeff Johnston <jjohnstn@redhat.com>
IP-Clean: Jeff Johnston <jjohnstn@redhat.com>
Tested-by: Jeff Johnston <jjohnstn@redhat.com>