Remove seek(time) from CtfTmfTrace and use parent implementation
The implementation of CtfTmfTrace of seek by timestamp worked 
correctly to seek within the CTF trace. However, the returned
context contains an incorrect rank of always 0. 

This incorrect rank will cause a wrong selected event in the
TmfEventsTable if the trace is used directly and not through an
experiment.

The solution is to use the TmfTrace implementation of seek(time)
instead which is handling the rank correctly.


Change-Id: Ie28c88c8e3263cbcb42349bb033cc61f69634a72
Reviewed-on: https://git.eclipse.org/r/8719
Reviewed-by: Patrick Tasse <patrick.tasse@gmail.com>
IP-Clean: Patrick Tasse <patrick.tasse@gmail.com>
Tested-by: Patrick Tasse <patrick.tasse@gmail.com>
Tested-by: Alexandre Montplaisir <alexmonthy@voxpopuli.im>