First version of remote unified profile launcher framework.
For clarity, this patch contains only the common framework to
support the remote unified launchers. The remote unified
launchers will be added in separated patches.

The reasons why I duplicated some of the classes were:

- The remote unified launcher won't necessarily work just like
the local unified. Some launch options (like choosing which
remote framework to use, like RSE, Remotetools, git ...) can change
the way the remote unified work in the future. However I haven't
duplicated all classes - the remote framework shares a lot of
code with the local one.

- Most of the APIs from the local unified framework are so tight
with the classes and return types used in the local launchers that
it would require a lot of code to make it work with the remote
launchers and it would be error prone. I really didn't want to take
the risk to add bugs in working code and, as I mentioned above,
the remote framework can diverge from the local one when more
remote options and support will be added in the future.

Change-Id: Id934a4572477dc2cb5193ff144d92af843c94173
Reviewed-on: https://git.eclipse.org/r/9239
Tested-by: Hudson CI
Reviewed-by: Otavio Pontes <obusatto@linux.vnet.ibm.com>
IP-Clean: Otavio Pontes <obusatto@linux.vnet.ibm.com>
Tested-by: Otavio Pontes <obusatto@linux.vnet.ibm.com>