Make selecting graph type better.
The label is removed and the descriptions are shown as tooltips. The
buttons are not toggle buttons so one can see what graph type is
currently selected. It was totally weird to not see what is selected and
to have to select something so you can see it's description.

Change-Id: Ib71d0dd46c5755570feadc50a6f89f165e80825e
Reviewed-on: https://git.eclipse.org/r/9571
Tested-by: Hudson CI
Reviewed-by: Sami Wagiaalla <swagiaal@redhat.com>
IP-Clean: Sami Wagiaalla <swagiaal@redhat.com>
Tested-by: Sami Wagiaalla <swagiaal@redhat.com>