Use Traces folder trace element when creating experiment from editor
When an experiment is created from the TmfEventsEditor.init method from
a FileEditorInput (ie. on platform restart or on open from bookmark),
the trace element from the Traces folder should be used. This is
required so that any plug-in extension that stores a persistent property
on the trace resource can access it when an experiment is created this
way.

Change-Id: I321a2cf3af91a7c898eb9800ce7f03fbbf34b9e4
Reviewed-on: https://git.eclipse.org/r/9940
Tested-by: Hudson CI
Reviewed-by: Alexandre Montplaisir <alexmonthy@voxpopuli.im>
IP-Clean: Alexandre Montplaisir <alexmonthy@voxpopuli.im>
Reviewed-by: Patrick Tasse <patrick.tasse@gmail.com>