Merge entries with same TID in Control Flow view
Separate entries are no longer created when a thread changes its exec
name over time. A single entry is used, and its name is set to the last
exec name of the thread. If the thread is terminated and reused, a
separate entry will be created with a different name and birth time.

Change-Id: I97d06adf8c68f9f56594059bc51a38fa78e36842
Reviewed-on: https://git.eclipse.org/r/10049
Tested-by: Hudson CI
Reviewed-by: Alexandre Montplaisir <alexmonthy@voxpopuli.im>
IP-Clean: Alexandre Montplaisir <alexmonthy@voxpopuli.im>
Tested-by: Alexandre Montplaisir <alexmonthy@voxpopuli.im>
Reviewed-by: Patrick Tasse <patrick.tasse@gmail.com>