tmf: Add support for versioning state input plugins
As we have seen with the recent addition of WAIT_FOR_CPU and
WAIT_BLOCKED state, if the framework gets updated we need a
way to tell it to rebuild any existing (and now outdated)
state history file that can be present in the workspace.

This patch reuses the minor_version field which was present
in the history tree's file header, but wasn't used. We now
assign this int to represent the version of the specific
state input we're trying to build. If the version in the
framework and the version of the existing file don't match, it
will force a rebuild.

Since the initial value used is 0, it should keep working with
any currently-existing history files (until the framework's
handlers are modified).

Change-Id: Id321db30f2e124cf93c96205a0450cccad47a49f
Signed-off-by: Alexandre Montplaisir <alexmonthy@voxpopuli.im>
Reviewed-on: https://git.eclipse.org/r/10638
Tested-by: Hudson CI
Reviewed-by: Bernd Hufmann <bhufmann@gmail.com>
IP-Clean: Bernd Hufmann <bhufmann@gmail.com>
Tested-by: Bernd Hufmann <bhufmann@gmail.com>