AbstractSTAnnotatedSourceEditorInput becomes an IAnnotationEditorInput.
To achieve that the number of STContributedColumns is limited to one
(there has never been a code trying to show what happens with two). This
changes the GCov editor input to be normal FileStoreEditorInput that
implements one more interface. The former allows working with plain
FileStoreEditorInput objects when needed and to expose less of ours.

Change-Id: I51fd4cf27f1791a28934048e093708137ebf56dc
Reviewed-on: https://git.eclipse.org/r/11031
Reviewed-by: Camilo Bernal <cabernal@redhat.com>
IP-Clean: Camilo Bernal <cabernal@redhat.com>
Tested-by: Camilo Bernal <cabernal@redhat.com>