tmf: Only keep the full constructor in TmfEventField
It's relatively risky to have two methods (in this case constructors)
where one uses 'Object' and the other 'ITmfEventField[]', since
an array is also an Object, so one constructor overloads the other.

We've had problems in the past with CtfTmfEventField, and it was
changed to only keep one constructor with all three arguments.
Propagate this change up to TmfEventField, and fixes all callers.
It's now much clearer what type of field we are constructing.

Change-Id: I3c19b9fffb53905f228f71b67200400c1ccc43ae
Signed-off-by: Alexandre Montplaisir <alexmonthy@voxpopuli.im>
Reviewed-on: https://git.eclipse.org/r/11734