tmf: Fix overwrite problem of batch import wizard
During batch import, when copying a trace, that is already in the
project as a link to an external location, with the same trace of
this external location then it used to wipe out the trace files in
the external location. This only happens when the trace is a
directory like the LTTng Kernel traces.
This commit fixes this problem.

Cherry-pick of 656357c7a0c2b73c39379f6d248c2296f430803c

Change-Id: I227f1eb07adb0b83699d8436f209badd0605770c
Signed-off-by: Bernd Hufmann <Bernd.Hufmann@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/13624
Tested-by: Hudson CI
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
IP-Clean: Matthew Khouzam <matthew.khouzam@ericsson.com>
Tested-by: Matthew Khouzam <matthew.khouzam@ericsson.com>