lttng rcp: add tracing rcp "file->open" buttons
there are two buttons added:
* File-> Open Trace File...
* File-> Open Trace Directory...

The traces will be opened into the default directory and
automatically select the trace type. If the trace viewer is
ambiguous, it will pop-up a window to help.

The trace is then opened.

The reason that we have open file and open directory is since
some traces are directories.

Also, the code path of regular trace file opening has been re-routed
through this patch to unify code.

Change-Id: I28df30745bf47b64ce0ab0f15a8d425fef4dd91a
Signed-off-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/13863
Reviewed-by: Alexandre Montplaisir <alexmonthy@voxpopuli.im>
IP-Clean: Alexandre Montplaisir <alexmonthy@voxpopuli.im>
Tested-by: Alexandre Montplaisir <alexmonthy@voxpopuli.im>