Systemtap: Prevent running scripts if existing graphs are missing required columns.
In Run Configurations->Graphing, deleting a parsing column that a created graph
relies on now counts as an error in a Run Configuration, and UI notifies users of
such an error when it is present.

A regular expression without any parenthesized groups also now counts as an error,
and graphs can't be added unless at least one such group exists.

Change-Id: Ibe6d11fad66c8b60ea40cd4faccb7cdf7749d18a
Signed-off-by: Andrew Ferrazzutti <aferrazz@redhat.com>
Reviewed-on: https://git.eclipse.org/r/16613
Tested-by: Hudson CI
Reviewed-by: Jeff Johnston <jjohnstn@redhat.com>
IP-Clean: Jeff Johnston <jjohnstn@redhat.com>
Tested-by: Jeff Johnston <jjohnstn@redhat.com>