tmf: Fix fake event problem in AbstractTmfStateProvider
For a given state provider, if the last event modifies the state
AND the user calls .waitForEmptyQueue() right before closing the
tree manually, the resulting history would have an incorrect end
time and it would be impossible to query at the timestamp of the
last event.

This was due do a "fake" synchronization event being set as the
provider's current event, so when closing the history it would
use its time stamp which is effectively 0.

Remedy this by using an internal event class explicitely for this.
Thanks to J.-C. Kouamé for finding this one!

Change-Id: I06227ca656a67991966b1d1dd15330cb97754fbe
Signed-off-by: Alexandre Montplaisir <alexmonthy@voxpopuli.im>
Reviewed-on: https://git.eclipse.org/r/14080
Tested-by: Hudson CI
Reviewed-by: Bernd Hufmann <bernd.hufmann@ericsson.com>