tmf: Consider compilation warnings to be errors in tmf.core
Now that we have eliminated all compiler warnings in the code,
we could turn on "-Werror" to make sure no new warnings are
introduced (Hudson CI would fail building any Gerrit patch that
has warnings).

NLS warnings remain as warnings, so that they can be suppressed
in Messages or constants classes.

"rawtype" and "unchecked" warnings for generic types remain as
warnings. Fixing some of those would require an API break.
Additionally, one of those warnings would require a major refactoring
of the ITmfLocation's, so this might remain as-is for the forseeable
future.

Finally, "Unecessary thrown exception" had to remain as a warning
because of TmfTrace.buildStateSystem(). The fix would require an API
break (but this is already fixed in the 3.0 branch, see commit e27acc1).

Change-Id: I810762a82ed4d0da2241ae943b1da1c2e96888d0
Signed-off-by: Alexandre Montplaisir <alexmonthy@voxpopuli.im>
Reviewed-on: https://git.eclipse.org/r/14998
Tested-by: Hudson CI
Reviewed-by: Bernd Hufmann <bernd.hufmann@ericsson.com>
IP-Clean: Bernd Hufmann <bernd.hufmann@ericsson.com>
Tested-by: Bernd Hufmann <bernd.hufmann@ericsson.com>