tmf: Fix Call Stack view event list
With the fix to the event iterator (42acdc2), it is no longer required
to add 'null' events in the zoomed event list to hide underlying events
in the full range event list. This simplifies the building of event list
in the Call Stack view.

If required, 'null' events are added at the start or end of the list to
properly identify the time range that the list represents.

Change-Id: I90ab8f91cc6991472a9ea5196a8e5cbee5a55565
Signed-off-by: Patrick Tasse <patrick.tasse@gmail.com>
Reviewed-on: https://git.eclipse.org/r/14599
Tested-by: Hudson CI