lttng: Disable API checking in the luna branch
Instead of always faking @since annotations to 3.0 in the luna branch,
it was deemed much more simple to simply ignore all API problems in
the Luna branch. That way all @since 2.x annotation will remain as-is,
without throwing errors when one checkouts that branch.

Instead of having to revert dozens and dozens of temporary commits when
we utlimately merge this branch back into master, only this one will
have to be reverted.

Change-Id: Ic37543462b677357b2051326d273d8504086db54
Signed-off-by: Alexandre Montplaisir <alexmonthy@voxpopuli.im>
Reviewed-on: https://git.eclipse.org/r/15553
Reviewed-by: Bernd Hufmann <bernd.hufmann@ericsson.com>
IP-Clean: Bernd Hufmann <bernd.hufmann@ericsson.com>