tmf: Update TmfSchedulerBenchmark requests
With ff830df currently sitting in the luna branch, the constructor
that was called by the inner request classes for the benchmark
does not exist anymore. However, turns out that its signature
matches the one of the new unique constructor. This had the nasty
effect of compiling fine, but giving completely meaningless results.

Update the constructor calls so that they use the new, relevant
parameters.

Change-Id: I5e86b40d735185fddb633cfedb754080fae09986
Signed-off-by: Alexandre Montplaisir <alexmonthy@voxpopuli.im>
Reviewed-on: https://git.eclipse.org/r/16145