RPM: modify behaviour of hyperlinks B#416446
The options available when Ctrl + hovering on any kind
of valid hyperlink within a specfile editor are modified based
on whether or not the hyperlink is a valid URL or file name.

Valid URLs will be given the option to OPEN IN A BROWSER. Valid
URLs that are linking to a file will be given the option to DOWNLOAD
the file. If the file does not exist and is a patch, it will offer to
CREATE it. If the file exists, then it will offer to OPEN the file.

Non URLs will either OPEN the file if it exists, or CREATE the file
if it is a patch.

Also, "Unkown Hyperlink" showing up when hovering over a macro and valid
URL will now show a simple text saying what it will do.

Bug: 416446

Change-Id: I51021ca1163dded22a74874e9133ff3c1b4f1b60
Signed-off-by: Neil Guzman <nguzman@redhat.com>
Reviewed-on: https://git.eclipse.org/r/16119
Tested-by: Hudson CI
Reviewed-by: Alexander Kurtakov <akurtako@redhat.com>
IP-Clean: Alexander Kurtakov <akurtako@redhat.com>
Tested-by: Alexander Kurtakov <akurtako@redhat.com>