systemtap: Use system colors where possible.
If one creates new Color it needs to be disposed manually which is not
an option with these static fields. For now just replace with system
color which are properly disposed by SWT itself. Not fixing the issue
entirely but at least minimizes it.

Change-Id: I0f9fe214df1c5d8f14f8e471e4ffc029f0c36583
Reviewed-on: https://git.eclipse.org/r/16946
Tested-by: Hudson CI
Reviewed-by: Alexander Kurtakov <akurtako@redhat.com>
IP-Clean: Alexander Kurtakov <akurtako@redhat.com>
Tested-by: Alexander Kurtakov <akurtako@redhat.com>