tmf: Optimize the ETERNITY time range
TmfTimeRange.ETERNITY is often used in requests and the like, when
you do not want to filter on timestamps at all. Implement it as
a subclass which does not even bother checking the range borders,
since any other timestamp or time range is definitely included.

Change-Id: I45c80523b054c33bd96f859355b795be38f7e760
Signed-off-by: Alexandre Montplaisir <alexmonthy@voxpopuli.im>
Reviewed-on: https://git.eclipse.org/r/17981
Tested-by: Hudson CI
Reviewed-by: Patrick Tasse <patrick.tasse@gmail.com>
IP-Clean: Patrick Tasse <patrick.tasse@gmail.com>