tmf: Add a debugging test Suite
When a test suite is run with the new DebugSuite runner, the name
of every test being run will be printed to stdout. It's not used
by default, but it can be manually switched to. This can help debug
unit tests, especially when run from Maven.

Change-Id: I5b3ca8a0a0f7867977500e5b3a280c46d702672c
Signed-off-by: Alexandre Montplaisir <alexmonthy@voxpopuli.im>
Reviewed-on: https://git.eclipse.org/r/18464
Tested-by: Hudson CI
Reviewed-by: Marc-Andre Laperle <marc-andre.laperle@ericsson.com>
IP-Clean: Marc-Andre Laperle <marc-andre.laperle@ericsson.com>
Tested-by: Marc-Andre Laperle <marc-andre.laperle@ericsson.com>